[test_nn] split pruning tests from test_nn
